requirements.txt
langchain>=0.1.0
openai>=1.0.0
milvus-client>=2.2.0
pinecone-client>=2.2.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
